api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.6.1
pinecone-client==3.0.0
pydantic==2.5.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
boto3==1.34.14
mangum==0.17.0
httpx==0.25.2
aiosqlite==0.19.0